prettier. imports

+++ contracts/StandardBondingCalculator.sol
@@ -4,7 +4,11 @@ pragma solidity 0.7.5;
 import "./SafeMath.sol";
 
 library FullMath {
-    function fullMul(uint256 x, uint256 y) private pure returns (uint256 l, uint256 h) {
+    function fullMul(uint256 x, uint256 y)
+        private
+        pure
+        returns (uint256 l, uint256 h)
+    {
         uint256 mm = mulmod(x, y, uint256(-1));
         l = x * y;
         h = mm - l;
@@ -41,13 +45,12 @@ library FullMath {
         uint256 mm = mulmod(x, y, d);
         if (mm > l) h -= 1;
         l -= mm;
-        require(h < d, 'FullMath::mulDiv: overflow');
+        require(h < d, "FullMath::mulDiv: overflow");
         return fullDiv(l, h, d);
     }
 }
 
 library Babylonian {
-
     function sqrt(uint256 x) internal pure returns (uint256) {
         if (x == 0) return 0;
 
@@ -93,9 +96,8 @@ library Babylonian {
 }
 
 library BitMath {
-
     function mostSignificantBit(uint256 x) internal pure returns (uint8 r) {
-        require(x > 0, 'BitMath::mostSignificantBit: zero');
+        require(x > 0, "BitMath::mostSignificantBit: zero");
 
         if (x >= 0x100000000000000000000000000000000) {
             x >>= 128;
@@ -144,7 +146,8 @@ library FixedPoint {
 
     uint8 private constant RESOLUTION = 112;
     uint256 private constant Q112 = 0x10000000000000000000000000000;
-    uint256 private constant Q224 = 0x100000000000000000000000000000000000000000000000000000000;
+    uint256 private constant Q224 =
+        0x100000000000000000000000000000000000000000000000000000000;
     uint256 private constant LOWER_MASK = 0xffffffffffffffffffffffffffff; // decimal of UQ*x112 (lower 112 bits)
 
     // decode a UQ112x112 into a uint112 by truncating after the radix point
@@ -153,35 +156,53 @@ library FixedPoint {
     }
 
     // decode a uq112x112 into a uint with 18 decimals of precision
-    function decode112with18(uq112x112 memory self) internal pure returns (uint) {
-        return uint(self._x) / 5192296858534827;
+    function decode112with18(uq112x112 memory self)
+        internal
+        pure
+        returns (uint256)
+    {
+        return uint256(self._x) / 5192296858534827;
     }
 
-    function fraction(uint256 numerator, uint256 denominator) internal pure returns (uq112x112 memory) {
-        require(denominator > 0, 'FixedPoint::fraction: division by zero');
+    function fraction(uint256 numerator, uint256 denominator)
+        internal
+        pure
+        returns (uq112x112 memory)
+    {
+        require(denominator > 0, "FixedPoint::fraction: division by zero");
         if (numerator == 0) return FixedPoint.uq112x112(0);
 
         if (numerator <= uint144(-1)) {
             uint256 result = (numerator << RESOLUTION) / denominator;
-            require(result <= uint224(-1), 'FixedPoint::fraction: overflow');
+            require(result <= uint224(-1), "FixedPoint::fraction: overflow");
             return uq112x112(uint224(result));
         } else {
             uint256 result = FullMath.mulDiv(numerator, Q112, denominator);
-            require(result <= uint224(-1), 'FixedPoint::fraction: overflow');
+            require(result <= uint224(-1), "FixedPoint::fraction: overflow");
             return uq112x112(uint224(result));
         }
     }
 
     // square root of a UQ112x112
     // lossy between 0/1 and 40 bits
-    function sqrt(uq112x112 memory self) internal pure returns (uq112x112 memory) {
+    function sqrt(uq112x112 memory self)
+        internal
+        pure
+        returns (uq112x112 memory)
+    {
         if (self._x <= uint144(-1)) {
             return uq112x112(uint224(Babylonian.sqrt(uint256(self._x) << 112)));
         }
 
         uint8 safeShiftBits = 255 - BitMath.mostSignificantBit(self._x);
         safeShiftBits -= safeShiftBits % 2;
-        return uq112x112(uint224(Babylonian.sqrt(uint256(self._x) << safeShiftBits) << ((112 - safeShiftBits) / 2)));
+        return
+            uq112x112(
+                uint224(
+                    Babylonian.sqrt(uint256(self._x) << safeShiftBits) <<
+                        ((112 - safeShiftBits) / 2)
+                )
+            );
     }
 }
 
@@ -190,61 +211,84 @@ interface IERC20 {
 }
 
 interface IUniswapV2ERC20 {
-    function totalSupply() external view returns (uint);
+    function totalSupply() external view returns (uint256);
 }
 
 interface IUniswapV2Pair is IUniswapV2ERC20 {
-    function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
-    function token0() external view returns ( address );
-    function token1() external view returns ( address );
+    function getReserves()
+        external
+        view
+        returns (
+            uint112 reserve0,
+            uint112 reserve1,
+            uint32 blockTimestampLast
+        );
+
+    function token0() external view returns (address);
+
+    function token1() external view returns (address);
 }
 
 interface IBondingCalculator {
-  function valuation( address pair_, uint amount_ ) external view returns ( uint _value );
+    function valuation(address pair_, uint256 amount_)
+        external
+        view
+        returns (uint256 _value);
 }
 
 contract OlympusBondingCalculator is IBondingCalculator {
-
     using FixedPoint for *;
-    using SafeMath for uint;
+    using SafeMath for uint256;
     using SafeMath for uint112;
 
     address public immutable OHM;
 
-    constructor( address _OHM ) {
-        require( _OHM != address(0) );
+    constructor(address _OHM) {
+        require(_OHM != address(0));
         OHM = _OHM;
     }
 
-    function getKValue( address _pair ) public view returns( uint k_ ) {
-        uint token0 = IERC20( IUniswapV2Pair( _pair ).token0() ).decimals();
-        uint token1 = IERC20( IUniswapV2Pair( _pair ).token1() ).decimals();
-        uint decimals = token0.add( token1 ).sub( IERC20( _pair ).decimals() );
+    function getKValue(address _pair) public view returns (uint256 k_) {
+        uint256 token0 = IERC20(IUniswapV2Pair(_pair).token0()).decimals();
+        uint256 token1 = IERC20(IUniswapV2Pair(_pair).token1()).decimals();
+        uint256 decimals = token0.add(token1).sub(IERC20(_pair).decimals());
 
-        (uint reserve0, uint reserve1, ) = IUniswapV2Pair( _pair ).getReserves();
-        k_ = reserve0.mul(reserve1).div( 10 ** decimals );
+        (uint256 reserve0, uint256 reserve1, ) = IUniswapV2Pair(_pair)
+            .getReserves();
+        k_ = reserve0.mul(reserve1).div(10**decimals);
     }
 
-    function getTotalValue( address _pair ) public view returns ( uint _value ) {
-        _value = getKValue( _pair ).sqrrt().mul(2);
+    function getTotalValue(address _pair) public view returns (uint256 _value) {
+        _value = getKValue(_pair).sqrrt().mul(2);
     }
 
-    function valuation( address _pair, uint amount_ ) external view override returns ( uint _value ) {
-        uint totalValue = getTotalValue( _pair );
-        uint totalSupply = IUniswapV2Pair( _pair ).totalSupply();
+    function valuation(address _pair, uint256 amount_)
+        external
+        view
+        override
+        returns (uint256 _value)
+    {
+        uint256 totalValue = getTotalValue(_pair);
+        uint256 totalSupply = IUniswapV2Pair(_pair).totalSupply();
 
-        _value = totalValue.mul( FixedPoint.fraction( amount_, totalSupply ).decode112with18() ).div( 1e18 );
+        _value = totalValue
+            .mul(FixedPoint.fraction(amount_, totalSupply).decode112with18())
+            .div(1e18);
     }
 
-    function markdown( address _pair ) external view returns ( uint ) {
-        ( uint reserve0, uint reserve1, ) = IUniswapV2Pair( _pair ).getReserves();
+    function markdown(address _pair) external view returns (uint256) {
+        (uint256 reserve0, uint256 reserve1, ) = IUniswapV2Pair(_pair)
+            .getReserves();
 
-        uint reserve;
-        if ( IUniswapV2Pair( _pair ).token0() == OHM ) {
+        uint256 reserve;
+        if (IUniswapV2Pair(_pair).token0() == OHM) {
             reserve = reserve1;
         } else {
             reserve = reserve0;
         }
-        return reserve.mul( 2 * ( 10 ** IERC20( OHM ).decimals() ) ).div( getTotalValue( _pair ) );
+        return
+            reserve.mul(2 * (10**IERC20(OHM).decimals())).div(
+                getTotalValue(_pair)
+            );
     }
 }
